+++ test/functional-tests/node-transformers/converting-transformers/split-string-transformer/fixtures/object-key-string-literal.js
@@ -1,0 +1,3 @@
+var test = {
+    'abcdefg': 1
+};